telegram_bot/requirements.txt
httpx==0.28.1
python-telegram-bot==21.9
